projectpage/requirements.txt
Django==1.6.2
Fabric==1.8.2
MySQL-python==1.2.4
PIL==1.1.7
Pillow==2.4.0
South==0.7.6
Whoosh==2.6.0
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.17
celery==3.1.11
coverage==3.7.1
distribute==0.6.35
django-allauth==0.16.1
-e git://github.com/amarandon/django-audiotracks.git@0b7d1d73da9df795e3d8927b2eec4326e4afb7cc#egg=django_audiotracks-dev
django-bootstrap-form==3.1
django-bootstrap-toolkit==2.15.0
-e git+https://github.com/django-debug-toolbar/django-debug-toolbar.git@568568c7225d506c6ab995410c1b1e45d95b5b52#egg=django_debug_toolbar-dev
django-google-tools==1.0.0
django-grappelli==2.5.2
django-haystack==2.1.0
django-hvad==0.3
django-pagination==1.0.7
django-user-media==0.11
ecdsa==0.11
gunicorn==18.0
kombu==3.0.16
mercurial==2.5.2
mutagen==1.22
mysql-connector-python==1.0.9
oauthlib==0.6.1
paramiko==1.13.0
phileo==1.1
psutil==2.0.0
psycopg2==2.4.6
pycrypto==2.6.1
python-openid==2.2.5
pytz==2014.2
requests==2.2.1
requests-oauthlib==0.4.0
sqlparse==0.1.11
stevedore==0.8
virtualenv==1.9.1
virtualenv-clone==0.2.4
virtualenvwrapper==3.6.1
wsgiref==0.1.2
yolk==0.4.3
